# Creating a blocking StatsDClient that is mostly based off of the existing NonBlockingStatsDClient