--- a/src/main/java/com/timgroup/statsd/StatsDClientErrorHandler.java
+++ b/src/main/java/com/timgroup/statsd/StatsDClientErrorHandler.java
@@ -16,4 +16,7 @@ public interface StatsDClientErrorHandler {
      */
     void handle(Exception exception);
 
+    StatsDClientErrorHandler NO_OP_HANDLER = new StatsDClientErrorHandler() {
+        @Override public void handle(final Exception e) { /* No-op */ }
+    };
 }