In a project I'm working on, it is preferred that we manage the asynchronous nature of sending to StatsD, so I made a blocking implementation that we can wrap and get the level of granularity we prefer.

In short:

1. I made a `BlockingStatsDClient` from the reusable parts of `NonBlockingStatsDClient`.
2. I made `NonBlockingStatsDClient` a subclass of `BlockingStatsDClient` and removed all common logic.
3. I made the test for `NonBlockingStatsDClient` generic so it can (and does) test both implementations.