@cowboygneox 2016-02-08
Also, it's interesting to note that I have made the tests faster (at least on my machine, so please verify).

Before:

```
-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Running com.timgroup.statsd.EventTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.029 sec
Running com.timgroup.statsd.NonBlockingStatsDClientPerfTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.338 sec
Running com.timgroup.statsd.NonBlockingStatsDClientTest
Tests run: 32, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 32.159 sec

Results :

Tests run: 38, Failures: 0, Errors: 0, Skipped: 0

[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time: 35.018 s
[INFO] Finished at: 2016-02-08T12:17:46-06:00
[INFO] Final Memory: 10M/245M
[INFO] ------------------------------------------------------------------------
```

After:

```
-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Running com.timgroup.statsd.BlockingStatsDClientTest
Tests run: 32, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.775 sec
Running com.timgroup.statsd.EventTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0 sec
Running com.timgroup.statsd.NonBlockingStatsDClientPerfTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.485 sec
Running com.timgroup.statsd.NonBlockingStatsDClientTest
Tests run: 32, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.885 sec

Results :

Tests run: 70, Failures: 0, Errors: 0, Skipped: 0

[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time: 5.625 s
[INFO] Finished at: 2016-02-08T12:19:05-06:00
[INFO] Final Memory: 10M/245M
[INFO] ------------------------------------------------------------------------
```