@cowboygneox 2016-04-12
@ovesh I reverted any/all changes to the `NonBlockingStatsDClient`. However, I did leave the tests in tact so that they would still cover both the blocking and non-blocking implementations. After reverting only `NonBlockingStatsDClient.java`, the tests still passed just fine, which is awesome.