@cowboygneox 2016-04-12
```-------------------------------------------------------
 T E S T S
-------------------------------------------------------
Running com.timgroup.statsd.BlockingStatsDClientTest
Tests run: 34, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 2.015 sec
Running com.timgroup.statsd.EventTest
Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.001 sec
Running com.timgroup.statsd.NonBlockingStatsDClientPerfTest
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 1.188 sec
Running com.timgroup.statsd.NonBlockingStatsDClientTest
Tests run: 34, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 34.175 sec

Results :

Tests run: 74, Failures: 0, Errors: 0, Skipped: 0

[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time: 39.744 s
[INFO] Finished at: 2016-04-11T20:53:35-05:00
[INFO] Final Memory: 10M/155M
[INFO] ------------------------------------------------------------------------```