@cowboygneox 2016-04-12
@ovesh While I'm fine with you taking this as-is, I would strongly advise that you consider reverting 	057a77b in the near future, since it should make your implementation more efficient. Might I suggest that you make a new major version so as not to break the reliability of the current version?