@cowboygneox 2016-04-12
I just realized that there are conflicts if you try to revert 057a77b. @ovesh if you are interested in improving the non-blocking client, let me know and I'll fix the conflicts.