@ovesh 2016-04-12
> if you are interested in improving the non-blocking client, let me know and I'll fix the conflicts.

Sure! Put that in a separate pull request.