@ovesh 2016-04-12
Sorry. I was going to merge and then I noticed 2 more things:

1. I think you missed some of the change from https://github.com/indeedeng/java-dogstatsd-client/pull/26, specifically https://github.com/cowboygneox/java-dogstatsd-client/blob/master/src/main/java/com/timgroup/statsd/BlockingStatsDClient.java#L172
2. Can you add a little documentation about the new blocking client to the readme?