@cowboygneox 2016-04-12
@ovesh I did a rebase and I'm pretty sure I captured what was necessary. Please review.