@cowboygneox 2016-04-12
@ovesh Also, even though I've spent a fair amount of time dinking around with this, I wonder if we don't just make a constructor for `QueueSize` for the blocking queue. I think that would accomplish both of our goals. Thoughts?