@cowboygneox 2016-04-12
@ovesh If you like this one better, take it instead and close this one: https://github.com/indeedeng/java-dogstatsd-client/pull/29