@ovesh 2016-04-14
A usecase just came up for the blocking client for us as well, so this PR is actually coming at a good time. Sorry to ask you for more work, but can you add support for a timeout? Things that may block forever are an antipattern.
If you feel like you've already spent too much time on this, I can take over from here.