@cowboygneox 2016-04-14
@ovesh I designed this pull request to allow the user to wrap in their own timeout logic. We actually wrap the blocking implementation in an actor that handles the management of timeouts.

I still think we should just take #29 and then just ensure that it will timeout when necessary.